refactor: Fixture 및 Fake 객체 패키지 이동
