+++ src/main/java/roomescape/web/config/LoginMemberArgumentResolver.java
@@ -40,6 +40,10 @@ public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolve
     }
 
     private String extractCookie(Cookie[] cookies, String targetCookie) {
+        if (cookies == null) {
+            throw new AuthenticationFailureException();
+        }
+
         return Arrays.stream(cookies)
                 .filter(cookie -> cookie.getName().equals(targetCookie))
                 .findAny()
