+++ src/test/java/roomescape/web/acceptance/AdminThemeControllerTest.java
@@ -1,8 +1,7 @@
-package roomescape.web.api;
+package roomescape.web.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
-import java.time.LocalTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -12,22 +11,21 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.context.TestExecutionListeners;
 import org.springframework.test.context.support.DependencyInjectionTestExecutionListener;
-import roomescape.application.dto.request.time.ReservationTimeRequest;
+import roomescape.application.dto.request.theme.ThemeRequest;
 import roomescape.application.security.JwtProvider;
 import roomescape.domain.member.Member;
-import roomescape.fixture.MemberFixture;
-import roomescape.fixture.TimeFixture;
 import roomescape.infrastructure.repository.MemberRepository;
-import roomescape.infrastructure.repository.ReservationTimeRepository;
+import roomescape.infrastructure.repository.ThemeRepository;
 import roomescape.support.DatabaseCleanupListener;
-
+import roomescape.support.fixture.MemberFixture;
+import roomescape.support.fixture.ThemeFixture;
 
 @TestExecutionListeners(value = {
         DatabaseCleanupListener.class,
         DependencyInjectionTestExecutionListener.class
 })
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-class AdminReservationTimeControllerTest {
+class AdminThemeControllerTest {
 
     @LocalServerPort
     private int port;
@@ -36,7 +34,7 @@ class AdminReservationTimeControllerTest {
     private JwtProvider jwtProvider;
 
     @Autowired
-    private ReservationTimeRepository timeRepository;
+    private ThemeRepository themeRepository;
 
     @Autowired
     private MemberRepository memberRepository;
@@ -50,28 +48,28 @@ class AdminReservationTimeControllerTest {
         adminToken = jwtProvider.encode(member);
     }
 
-    @DisplayName("예약 시간을 생성하는데 성공하면 응답과 201 상태 코드를 반환한다.")
+    @DisplayName("테마를 생성하는데 성공하면 응답과 201 상태 코드를 반환한다.")
     @Test
-    void return_201_when_create_reservation_time() {
-        ReservationTimeRequest request = new ReservationTimeRequest(LocalTime.parse("10:00"));
+    void return_201_when_create_theme() {
+        ThemeRequest request = new ThemeRequest("테마다", "설명이다", "썸네일이다");
 
         RestAssured.given().log().all()
                 .cookie("token", adminToken)
                 .contentType(ContentType.JSON)
                 .body(request)
-                .when().post("/admin/times")
+                .when().post("/admin/themes")
                 .then().log().all()
                 .statusCode(201);
     }
 
-    @DisplayName("예약 시간을 삭제하는데 성공하면 응답과 204 상태 코드를 반환한다.")
+    @DisplayName("테마를 삭제하는데 성공하면 응답과 204 상태 코드를 반환한다.")
     @Test
-    void return_204_when_delete_reservation_time() {
-        timeRepository.save(TimeFixture.ONE_PM.create());
+    void return_204_when_delete_theme() {
+        themeRepository.save(ThemeFixture.THEME_JAVA.create());
 
         RestAssured.given().log().all()
                 .cookie("token", adminToken)
-                .when().delete("/admin/times/1")
+                .when().delete("/admin/themes/1")
                 .then().log().all()
                 .statusCode(204);
     }
