+++ src/test/java/roomescape/web/acceptance/MemberControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.web.api;
+package roomescape.web.acceptance;
 
 import static org.hamcrest.Matchers.is;
 
@@ -13,19 +13,18 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.context.TestExecutionListeners;
 import org.springframework.test.context.support.DependencyInjectionTestExecutionListener;
-import roomescape.application.dto.request.member.SignupRequest;
 import roomescape.application.security.JwtProvider;
 import roomescape.domain.member.Member;
-import roomescape.fixture.MemberFixture;
 import roomescape.infrastructure.repository.MemberRepository;
 import roomescape.support.DatabaseCleanupListener;
+import roomescape.support.fixture.MemberFixture;
 
 @TestExecutionListeners(value = {
         DatabaseCleanupListener.class,
         DependencyInjectionTestExecutionListener.class
 })
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-class MemberManagementControllerTest {
+class MemberControllerTest {
 
     @LocalServerPort
     private int port;
@@ -36,26 +35,13 @@ class MemberManagementControllerTest {
     @Autowired
     private MemberRepository memberRepository;
 
-    String adminToken;
+    private String adminToken;
 
     @BeforeEach
     void setUp() {
         RestAssured.port = port;
         Member member = memberRepository.save(MemberFixture.MEMBER_SOLAR.create());
         adminToken = jwtProvider.encode(member);
-    }
-
-    @DisplayName("회원가입에 성공하면 응답과 201 상태코드를 반환한다.")
-    @Test
-    void return_201_when_signup() {
-        SignupRequest request = new SignupRequest("재즈", "jazz@woowa.com", "123");
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(request)
-                .when().post("/members")
-                .then().log().all()
-                .statusCode(201);
     }
 
     @DisplayName("멤버 목록을 조회하는데 성공하면 응답과 200 상태코드를 반환한다.")
@@ -77,7 +63,7 @@ class MemberManagementControllerTest {
 
     @DisplayName("멤버를 삭제하는데 성공하면 응답과 204 상태코드를 반환한다.")
     @Test
-    void return_200_when_delete_member() {
+    void return_204_when_delete_member() {
         memberRepository.save(MemberFixture.MEMBER_BRI.create());
 
         RestAssured.given()
